+++ packages/core/src/parsers/builder/builder.ts
@@ -353,6 +353,53 @@ const componentMappers: {
 
     return node;
   },
+  PersonalizationContainer(block, options) {
+    const node = builderElementToMitosisNode(block, options, {
+      skipMapper: true,
+    });
+
+    delete node.bindings.variants;
+    delete node.properties.variants;
+
+    const newChildren: MitosisNode[] =
+      block.component?.options.variants?.map((variant: any) => {
+        const variantNode = createMitosisNode({
+          name: 'Variant',
+          properties: {
+            name: variant.name,
+            startDate: variant.startDate,
+            endDate: variant.endDate,
+          },
+          meta: getMetaFromBlock(block, options),
+          children: variant.blocks.map((col: any) => builderElementToMitosisNode(col, options)),
+        });
+        const queryOptions = variant.query as any[];
+        if (Array.isArray(queryOptions)) {
+          variantNode.bindings.query = {
+            type: 'single',
+            code: JSON.stringify(queryOptions.map((q) => omit(q, '@type'))),
+          };
+        } else if (queryOptions) {
+          variantNode.bindings.query = {
+            type: 'single',
+            code: JSON.stringify(omit(queryOptions, '@type')),
+          };
+        }
+        return variantNode;
+      }) || [];
+
+    const defaultVariant = createMitosisNode({
+      name: 'Variant',
+      properties: {
+        default: '',
+      },
+      children: node.children,
+    });
+    newChildren.push(defaultVariant);
+
+    node.children = newChildren;
+    return node;
+  },
   'Shopify:For': (block, options) => {
     return createMitosisNode({
       name: 'For',
@@ -417,10 +464,14 @@ const componentMappers: {
     }
     const text = block.component!.options.text;
 
+    // Builder uses {{}} for bindings, but Mitosis expects {} so we need to convert
     const innerProperties = innerBindings._text
       ? {}
       : {
-          [options.preserveTextBlocks ? 'innerHTML' : '_text']: text,
+          [options.preserveTextBlocks ? 'innerHTML' : '_text']: text.replace(
+            /\{\{(.*?)\}\}/g,
+            '{$1}',
+          ),
         };
 
     if (options.preserveTextBlocks) {
@@ -1014,8 +1065,10 @@ const builderContentPartToMitosisComponent = (
       useMetadata: {
         httpRequests: builderContent.data?.httpRequests,
       },
+      // cmp.meta.cssCode exists for backwards compatibility, prefer cmp.style
       ...(builderContent.data?.cssCode && { cssCode: builderContent.data.cssCode }),
     },
+    ...(builderContent.data?.cssCode && { style: builderContent.data?.cssCode }),
     inputs: builderContent.data?.inputs?.map((input) => ({
       name: input.name,
       defaultValue: input.defaultValue,